refactored stats functions
